+++ examples/create_react_app/dark-mode-app/src/main.jsx
@@ -27,7 +27,7 @@ export default function App() {
           Dark mode
         </label>
       </div>
-      <ReactaForm definitionData={{ ...darkModeDefinition }} darkMode={darkMode} />
+      <ReactaForm definitionData={{ ...darkModeDefinition }} theme={darkMode ? 'dark' : 'light'} />
     </div>
   );
 }